Move velocity related files to separate module
