+++ settings.gradle
@@ -3,3 +3,5 @@ include 'api'
 include 'spigot-legacy'
 include 'v1_13'
 include 'core'
+include 'triton-velocity'
+
