+++ triton-velocity/build.gradle
@@ -1,0 +1,12 @@
+plugins {
+    id 'java'
+}
+
+group 'com.rexcantor64.triton'
+
+dependencies {
+    compileOnly project(path: ":core")
+
+    compileOnly 'com.velocitypowered:velocity-api:3.1.0'
+    annotationProcessor 'com.velocitypowered:velocity-api:3.1.0'
+}
