+++ triton-velocity/src/main/java/com/rexcantor64/triton/velocity/plugin/VelocityPlugin.java
@@ -1,9 +1,10 @@
-package com.rexcantor64.triton.plugin;
+package com.rexcantor64.triton.velocity.plugin;
 
 import com.google.inject.Inject;
 import com.rexcantor64.triton.VelocityMLP;
 import com.rexcantor64.triton.logger.SLF4JLogger;
 import com.rexcantor64.triton.logger.TritonLogger;
+import com.rexcantor64.triton.plugin.PluginLoader;
 import com.velocitypowered.api.event.Subscribe;
 import com.velocitypowered.api.event.proxy.ProxyInitializeEvent;
 import com.velocitypowered.api.plugin.Plugin;
